% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ten.R
\name{add_ten}
\alias{add_ten}
\title{add_ten}
\usage{
add_ten(number)
}
\arguments{
\item{number}{}
}
\value{
number + 10
}
\description{
Add 10 to a given number
}
\examples{
add_ten(22) will return 32
add_ten(1:3) will return 11 12 13
}
