% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minus_ten.R
\name{minus_ten}
\alias{minus_ten}
\title{Minus Ten}
\usage{
minus_ten(number)
}
\arguments{
\item{number}{a number}
}
\value{
a number that is the provided number minus 10
}
\description{
Subtract 10 from a given number
}
\examples{
minus_ten(22) will return 12
minus_ten(22:25) will return 12 13 14 15
}
