% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{randomNumbers}
\alias{randomNumbers}
\title{randomNumbers}
\format{A vector with 200 numbers}
\usage{
randomNumbers
}
\description{
200 random numbers generated as the output of sample(200)
}
\keyword{datasets}
